fix 7 log

+++ gazebo_plugins/include/gazebo_plugins/gazebo_ros_actuator.h
@@ -45,7 +45,7 @@ namespace gazebo {
         std::string joint_name_;
 
         double torque;
-        float64 position_;
+        float position_;
 
         physics::JointPtr joints;
